Polymorphic len as a library function rather than primitive.
The list of primitives was originally intended to be used for things
that can't be expressed in the source language (usually pertaining to
the internal representation of distributions, so for example
primitives would be defined for constructing trees in particular
ways). They seemingly can't be made polymorphic with the current setup
because their types must be recorded in the list of primitives once
and for all (their must be an object language level type that
completely captures their type), and we can't support forall types in
GADTs AFAIK.

However, for "library" functions that don't care about the particular
representation being used (and therefore needn't be resolved by their
association to the chosen 'm' and 'g'), we can just define them as
polymorphic terms in Haskell without being constrained by the need to
record their type within the object language type system. So basically
we have access, at the level of Haskell, to an infinite family of
monomorphic list functions in the object language.

I *think* this will work for these polymorphic list functions at
least, but not totally sure.