Fix to support exact inference on the program in halfmass.zar.
Currently only the custom solver in LinEq.hs is fixed for this.

Also slightly refactored the Tree interpretation of commands as
kleisli arrows out of St instead of tree transformers. That is, a `Com
a` is now interpreted as a function of type `St -> InterpM (Tree a)`
instead of `Tree St -> InterpM (Tree a)`, and the interpretation of
Seq handles the dirty work of composition. Need to think about this
more because I'm not sure that it will enable what I want -- may end
up reverting.